@@ -22,28 +22,34 @@
 
   <properties>
     <!-- runtime dependencies -->
-    <folio-spring-base.version>5.0.3</folio-spring-base.version>
-    <spring-cloud-starter-openfeign.version>3.1.5</spring-cloud-starter-openfeign.version>
+    <folio-spring-base.version>6.0.0</folio-spring-base.version>
+    <spring-cloud-starter-openfeign.version>4.0.0</spring-cloud-starter-openfeign.version>
     <commons-collections4.version>4.4</commons-collections4.version>
-    <mapstruct.version>1.5.1.Final</mapstruct.version>
+    <mapstruct.version>1.5.3.Final</mapstruct.version>
     <marc4j.version>2.9.2</marc4j.version>
     <coffee-boots.version>3.0.0</coffee-boots.version>
     <snakeyaml.version>1.33</snakeyaml.version>
 
     <!-- test dependencies -->
-    <testcontainers.version>1.17.2</testcontainers.version>
-    <wiremock.version>2.33.2</wiremock.version>
+    <testcontainers.version>1.17.6</testcontainers.version>
+    <wiremock.version>2.35.0</wiremock.version>
     <awaitility.version>4.2.0</awaitility.version>
-    <junit-extensions.version>2.5.0</junit-extensions.version>
+    <junit-extensions.version>2.6.0</junit-extensions.version>
     <easy-random.version>5.0.0</easy-random.version>
-    <micrometer.version>1.9.4</micrometer.version>
+    <micrometer.version>1.10.3</micrometer.version>
     <log4j2.version>2.19.0</log4j2.version>
+    <jetty.version>9.4.49.v20220914</jetty.version>
 
     <!-- plugins -->
-    <openapi-generator.version>4.3.1</openapi-generator.version>
+    <openapi-generator.version>6.2.1</openapi-generator.version>
+    <copy-rename-maven-plugin.version>1.0.1</copy-rename-maven-plugin.version>
+    <build-helper-maven-plugin.version>3.3.0</build-helper-maven-plugin.version>
+    <maven-surefire-plugin.version>3.0.0-M7</maven-surefire-plugin.version>
+    <maven-release-plugin.version>3.0.0-M7</maven-release-plugin.version>
+    <maven-enforcer-plugin.version>3.1.0</maven-enforcer-plugin.version>
 
     <!-- other properties -->
-    <java.version>11</java.version>
+    <java.version>17</java.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
     <openapi.settings.file>${project.basedir}/src/main/resources/swagger.api/settings.yaml</openapi.settings.file>
