@@ -92,7 +92,9 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.support.TransactionTemplate;
 import org.springframework.util.Assert;
 
-import javax.persistence.EntityExistsException;
+import jakarta.persistence.EntityExistsException;
+
+import java.time.Instant;
 import java.util.Collection;
 import java.util.Date;
 import java.util.HashMap;
@@ -488,7 +490,7 @@ public class CirculationServiceImpl implements CirculationService {
             () -> virtualRecordService.deleteVirtualRecords(folioItemId, folioHoldingId,
                     folioInstanceId, folioLoanId));
 
-    taskExecutor.schedule(task, new Date(System.currentTimeMillis() + checkOutTimeDuration));
+    taskExecutor.schedule(task, Instant.now().plusMillis(checkOutTimeDuration));
   }
 
   @Override
