@@ -2,12 +2,11 @@ package org.folio.innreach.domain.service.impl;
 
 import static org.folio.innreach.domain.service.impl.ServiceUtils.centralServerRef;
 
+import jakarta.persistence.EntityExistsException;
+
 import java.util.Comparator;
 import java.util.List;
 import java.util.UUID;
-import java.util.stream.Collectors;
-
-import javax.persistence.EntityExistsException;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -92,7 +91,7 @@ public class MARCTransformationOptionsSettingsServiceImpl implements MARCTransfo
     return repository.findAll(new OffsetRequest(offset, limit))
       .stream()
       .map(mapper::toDto)
-      .collect(Collectors.toList());
+      .toList();
   }
 
   private void updateMARCTransformOptSet(MARCTransformationOptionsSettings marcTransformOptSet, MARCTransformationOptionsSettings updated) {
