@@ -8,13 +8,13 @@ import static org.folio.innreach.domain.entity.InnReachTransaction.TransactionTy
 import static org.folio.innreach.domain.entity.InnReachTransactionFilterParameters.SortOrder.DESC;
 import static org.folio.innreach.util.ListUtils.mapItems;
 
-import java.time.OffsetDateTime;
+import jakarta.persistence.criteria.CriteriaBuilder;
+import jakarta.persistence.criteria.Expression;
+import jakarta.persistence.criteria.Join;
+import jakarta.persistence.criteria.Predicate;
+import jakarta.persistence.criteria.Root;
 
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.Expression;
-import javax.persistence.criteria.Join;
-import javax.persistence.criteria.Predicate;
-import javax.persistence.criteria.Root;
+import java.time.OffsetDateTime;
 
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.data.jpa.domain.Specification;
@@ -24,7 +24,6 @@ import org.folio.innreach.domain.entity.InnReachTransaction;
 import org.folio.innreach.domain.entity.InnReachTransactionFilterParameters;
 import org.folio.innreach.domain.entity.InnReachTransactionFilterParameters.SortBy;
 import org.folio.innreach.domain.entity.InnReachTransactionFilterParameters.SortOrder;
-import org.folio.innreach.domain.entity.TransactionPatronHold;
 
 @Component
 public class InnReachTransactionSpecification {
@@ -45,7 +44,6 @@ public class InnReachTransactionSpecification {
     return (transaction, cq, cb) -> {
       var isRequestTooLongReport = parameters.isRequestedTooLong();
       var hold = transaction.join("hold");
-      var patronHold = cb.treat(hold, TransactionPatronHold.class);
 
       var typeIs = isOfType(cb, transaction, parameters);
       var stateIs = isOfState(cb, transaction, parameters);
@@ -55,7 +53,7 @@ public class InnReachTransactionSpecification {
       var patronTypeIn = patronTypeIn(cb, hold, parameters);
       var patronNameIn = patronNameIn(cb, hold, parameters);
       var centralItemTypeIn = centralItemTypeIn(cb, hold, parameters);
-      var itemBarcodeIn = itemBarcodeIn(cb, transaction, hold, patronHold, parameters);
+      var itemBarcodeIn = itemBarcodeIn(cb, transaction, hold, parameters);
 
       var odtConditionFactory = new ConditionFactory<OffsetDateTime>(cb);
 
@@ -171,7 +169,6 @@ public class InnReachTransactionSpecification {
   static Predicate itemBarcodeIn(CriteriaBuilder cb,
                                  Root<InnReachTransaction> transaction,
                                  Join<Object, Object> hold,
-                                 Join<Object, TransactionPatronHold> patronHold,
                                  InnReachTransactionFilterParameters parameters) {
     var itemBarcodes = parameters.getItemBarcodes();
     if (isEmpty(itemBarcodes)) {
@@ -180,7 +177,7 @@ public class InnReachTransactionSpecification {
 
     var shippedItemBarcodePredicate = cb.and(
       cb.equal(transaction.get("type"), PATRON),
-      patronHold.get("shippedItemBarcode").in(itemBarcodes)
+      hold.get("shippedItemBarcode").in(itemBarcodes)
     );
 
     var folioItemBarcodePredicate = cb.and(
@@ -208,18 +205,10 @@ public class InnReachTransactionSpecification {
   }
 
   private static Expression<InnReachTransaction> getField(Root<InnReachTransaction> root, SortBy sort) {
-    Expression<InnReachTransaction> expression;
-    switch (sort) {
-      case TRANSACTION_TYPE:
-      case TRANSACTION_STATUS:
-      case DATE_CREATED:
-      case DATE_MODIFIED:
-        expression = root.get(sort.getValue());
-        break;
-      default:
-        expression = root.get("hold").get(sort.getValue());
-    }
-    return expression;
+    return switch (sort) {
+      case TRANSACTION_TYPE, TRANSACTION_STATUS, DATE_CREATED, DATE_MODIFIED -> root.get(sort.getValue());
+      default -> root.get("hold").get(sort.getValue());
+    };
   }
 
 }
