@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.folio.innreach.client.config.FolioFeignClientConfig;
 import org.folio.innreach.dto.Item;
 
-@FeignClient(name = "item-storage", configuration = FolioFeignClientConfig.class, decode404 = true)
+@FeignClient(name = "item-storage", configuration = FolioFeignClientConfig.class, dismiss404 = true)
 public interface ItemStorageClient {
 
   @GetMapping("/items/{itemId}")
