@@ -16,7 +16,7 @@ import org.folio.innreach.client.config.FolioFeignClientConfig;
 import org.folio.innreach.domain.dto.folio.ResultList;
 import org.folio.innreach.domain.dto.folio.User;
 
-@FeignClient(name = "users", configuration = FolioFeignClientConfig.class, decode404 = true)
+@FeignClient(name = "users", configuration = FolioFeignClientConfig.class, dismiss404 = true)
 public interface UsersClient {
 
   @GetMapping(value = "/{id}")
