@@ -3,30 +3,28 @@ package org.folio.innreach.domain.entity;
 import static org.folio.innreach.domain.entity.AgencyLocationMapping.FETCH_ONE_BY_CS_QUERY;
 import static org.folio.innreach.domain.entity.AgencyLocationMapping.FETCH_ONE_BY_CS_QUERY_NAME;
 
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.NamedQuery;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.OrderBy;
+import jakarta.persistence.Table;
+
 import java.util.LinkedHashSet;
 import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.NamedQuery;
-import javax.persistence.OneToMany;
-import javax.persistence.OneToOne;
-import javax.persistence.OrderBy;
-import javax.persistence.QueryHint;
-import javax.persistence.Table;
-
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-import org.hibernate.annotations.QueryHints;
 
 import org.folio.innreach.domain.entity.base.Auditable;
 import org.folio.innreach.domain.entity.base.Identifiable;
@@ -38,8 +36,7 @@ import org.folio.innreach.domain.entity.base.Identifiable;
 @Entity
 @NamedQuery(
   name = FETCH_ONE_BY_CS_QUERY_NAME,
-  query = FETCH_ONE_BY_CS_QUERY,
-  hints = @QueryHint(name = QueryHints.PASS_DISTINCT_THROUGH, value = "false")
+  query = FETCH_ONE_BY_CS_QUERY
 )
 @Table(name = "agency_location_mapping")
 public class AgencyLocationMapping extends Auditable implements Identifiable<UUID> {
