@@ -17,30 +17,28 @@ import static org.folio.innreach.domain.entity.CentralServer.GET_IDS_QUERY_NAME;
 import static org.folio.innreach.domain.entity.CentralServer.GET_ID_BY_CENTRAL_CODE_QUERY;
 import static org.folio.innreach.domain.entity.CentralServer.GET_ID_BY_CENTRAL_CODE_QUERY_NAME;
 
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.NamedQuery;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
-
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.NamedQuery;
-import javax.persistence.OneToMany;
-import javax.persistence.OneToOne;
-import javax.persistence.QueryHint;
-import javax.persistence.Table;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-import org.hibernate.annotations.QueryHints;
 
 import org.folio.innreach.domain.entity.base.Auditable;
 import org.folio.innreach.domain.entity.base.Identifiable;
@@ -53,18 +51,15 @@ import org.folio.innreach.domain.entity.base.Identifiable;
 @Table(name = "central_server")
 @NamedQuery(
   name = FETCH_ALL_BY_ID_QUERY_NAME,
-  query = FETCH_ALL_BY_ID_QUERY,
-  hints = @QueryHint(name = QueryHints.PASS_DISTINCT_THROUGH, value = "false")
+  query = FETCH_ALL_BY_ID_QUERY
 )
 @NamedQuery(
   name = FETCH_ONE_BY_ID_QUERY_NAME,
-  query = FETCH_ONE_BY_ID_QUERY,
-  hints = @QueryHint(name = QueryHints.PASS_DISTINCT_THROUGH, value = "false")
+  query = FETCH_ONE_BY_ID_QUERY
 )
 @NamedQuery(
   name = FETCH_ONE_BY_CENTRAL_CODE_QUERY_NAME,
-  query = FETCH_ONE_BY_CENTRAL_CODE_QUERY,
-  hints = @QueryHint(name = QueryHints.PASS_DISTINCT_THROUGH, value = "false")
+  query = FETCH_ONE_BY_CENTRAL_CODE_QUERY
 )
 @NamedQuery(
   name = GET_IDS_QUERY_NAME,
@@ -76,13 +71,11 @@ import org.folio.innreach.domain.entity.base.Identifiable;
 )
 @NamedQuery(
   name = FETCH_CONNECTION_DETAILS_BY_ID_QUERY_NAME,
-  query = FETCH_CONNECTION_DETAILS_BY_ID_QUERY,
-  hints = @QueryHint(name = QueryHints.PASS_DISTINCT_THROUGH, value = "false")
+  query = FETCH_CONNECTION_DETAILS_BY_ID_QUERY
 )
 @NamedQuery(
   name = FETCH_CONNECTION_DETAILS_BY_CENTRAL_CODE_QUERY_NAME,
-  query = FETCH_CONNECTION_DETAILS_BY_CENTRAL_CODE_QUERY,
-  hints = @QueryHint(name = QueryHints.PASS_DISTINCT_THROUGH, value = "false")
+  query = FETCH_CONNECTION_DETAILS_BY_CENTRAL_CODE_QUERY
 )
 @NamedQuery(
   name = FETCH_RECALL_USER_BY_ID_QUERY_NAME,
