@@ -6,31 +6,29 @@ import static org.folio.innreach.domain.entity.Contribution.FETCH_HISTORY_COUNT_
 import static org.folio.innreach.domain.entity.Contribution.FETCH_HISTORY_QUERY;
 import static org.folio.innreach.domain.entity.Contribution.FETCH_HISTORY_QUERY_NAME;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.NamedQuery;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Enumerated;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.NamedQuery;
-import javax.persistence.OneToMany;
-import javax.persistence.OneToOne;
-import javax.persistence.QueryHint;
-import javax.persistence.Table;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-import org.hibernate.annotations.QueryHints;
 
 import org.folio.innreach.domain.entity.base.Auditable;
 import org.folio.innreach.domain.entity.base.Identifiable;
@@ -42,14 +40,12 @@ import org.folio.innreach.domain.entity.base.Identifiable;
 @Entity
 @NamedQuery(
   name = FETCH_HISTORY_QUERY_NAME,
-  query = FETCH_HISTORY_QUERY,
-  hints = @QueryHint(name = QueryHints.PASS_DISTINCT_THROUGH, value = "false")
+  query = FETCH_HISTORY_QUERY
 )
 @NamedQuery(name = FETCH_HISTORY_COUNT_QUERY_NAME, query = Contribution.FETCH_HISTORY_COUNT_QUERY)
 @NamedQuery(
   name = FETCH_CURRENT_QUERY_NAME,
-  query = FETCH_CURRENT_QUERY,
-  hints = @QueryHint(name = QueryHints.PASS_DISTINCT_THROUGH, value = "false")
+  query = FETCH_CURRENT_QUERY
 )
 @Table(name = "contribution")
 public class Contribution extends Auditable implements Identifiable<UUID> {
