@@ -5,32 +5,30 @@ import static org.hibernate.annotations.FetchMode.SUBSELECT;
 import static org.folio.innreach.domain.entity.VisiblePatronFieldConfiguration.FETCH_ONE_BY_CENTRAL_CODE_QUERY;
 import static org.folio.innreach.domain.entity.VisiblePatronFieldConfiguration.FETCH_ONE_BY_CENTRAL_CODE_QUERY_NAME;
 
+import jakarta.persistence.CollectionTable;
+import jakarta.persistence.Column;
+import jakarta.persistence.ElementCollection;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.NamedQuery;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+
 import java.util.ArrayList;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
 
-import javax.persistence.CollectionTable;
-import javax.persistence.Column;
-import javax.persistence.ElementCollection;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.NamedQuery;
-import javax.persistence.OneToOne;
-import javax.persistence.QueryHint;
-import javax.persistence.Table;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 import org.hibernate.annotations.Fetch;
-import org.hibernate.annotations.QueryHints;
 
 import org.folio.innreach.domain.entity.base.Auditable;
 import org.folio.innreach.domain.entity.base.Identifiable;
@@ -42,8 +40,7 @@ import org.folio.innreach.domain.entity.base.Identifiable;
 @ToString(exclude = {"centralServer"})
 @NamedQuery(
   name = FETCH_ONE_BY_CENTRAL_CODE_QUERY_NAME,
-  query = FETCH_ONE_BY_CENTRAL_CODE_QUERY,
-  hints = @QueryHint(name = QueryHints.PASS_DISTINCT_THROUGH, value = "false")
+  query = FETCH_ONE_BY_CENTRAL_CODE_QUERY
 )
 public class VisiblePatronFieldConfiguration extends Auditable implements Identifiable<UUID> {
   public static final String FETCH_ONE_BY_CENTRAL_CODE_QUERY_NAME = "VisiblePatronFieldConfiguration.fetchOneByCode";
